Make ReplaceIterator and Replacing Iterator works.
Move EI rules : suffixing and removing of level 01 into Replacing iterator.
ReplaceIterator now use the AbstractReplaceTokensLinesIterator.
ReplaceIterator2 is deleted.
