--- TypeCobol/Compiler/Preprocessor/CopyTokensLinesIterator2.cs
@@ -1,9 +1,0 @@
-﻿using TypeCobol.Compiler.Scanner;
-
-namespace TypeCobol.Compiler.Preprocessor
-{
-    public class CopyTokensLinesIterator2 : ITokensLinesIterator
-    {
-        // Same as current but includes legacy replacing mechanism
-    }
-}
